requirements.txt
Aspose.Slides==22.8.0
fastapi==0.85.0
numpy==1.23.3
opencv-python==4.6.0.66
pandas==1.5.0
pdf2docx==0.5.6
pdf2image==1.16.0
Pillow==9.2.0
pydantic==1.10.2
PyMuPDF==1.20.2
python-docx==0.8.11
python-multipart==0.0.5
starlette==0.20.4
tabula-py==2.5.1
uvicorn==0.18.3

mutagen~=1.45.1
moviepy~=1.0.3
pytesseract~=0.3.10
pdftotree~=0.5.0
PyPDF2~=2.11.0
PDFNetPython3~=9.3.0
docx2pdf~=0.1.8
tensorflow~=2.10.0
qrcode~=7.3.1
keras~=2.10.0